filling gaps between files

+++ dug_seis/waveform_handler/waveform_handler.py
@@ -575,8 +575,8 @@ class WaveformHandler:
                 gap_st = st.select(id=i).copy()
                 gaps = len(gap_st) - 1
                 for gap_no in range(gaps):
-                    samples = (st[gap_no + 1].stats.starttime -
-                               st[gap_no].stats.endtime) * st[gap_no].stats.sampling_rate
+                    samples = int((st[gap_no + 1].stats.starttime -
+                                   st[gap_no].stats.endtime) * st[gap_no].stats.sampling_rate)
                     mad = np.median(np.abs(st[gap_no].data - np.mean(st[gap_no].data)))
                     fill = np.linspace(st[gap_no].data[-1], st[gap_no + 1].data[0], samples)
                     # Fill with noise
